dyn_memdb/requirements.txt
asyncpg>=0.28.0
python-dotenv>=1.0.0